refactor: split up functionality in main.py for more modularity
