+++ src/globals.py
@@ -1,0 +1,7 @@
+from pathlib import Path
+
+# Build path relative to absolute path of script (or rather the scripts parent dir)
+# Source: https://stackoverflow.com/a/55051039
+BASE_PATH = Path(__file__).parent
+CREDENTIALS_PATH = (BASE_PATH / "./credentials.json").resolve()
+DOWNLOAD_CONFIG_PATH = (BASE_PATH / "./download_config.json").resolve()
